% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{averageDuplicates}
\alias{averageDuplicates}
\title{averageDuplicates}
\usage{
averageDuplicates(geneDataFrame)
}
\arguments{
\item{geneDataFrame}{Dataframe produced by readGenes.}
}
\value{
geneDataFrame with duplicates averaged.
}
\description{
This function takes a dataframe produced by readGenes.
}
