% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{barplotControlRelativizedGeneExpression}
\alias{barplotControlRelativizedGeneExpression}
\title{barplotControlRelativizedGeneExpression}
\usage{
barplotControlRelativizedGeneExpression(geneDataFrame, labels = F)
}
\arguments{
\item{geneDataFrame}{Dataframe produced by readGenes with averaged duplicates.}

\item{labels}{Whether to include sample labels.}
}
\description{
This function plots relative expression by group relativized to the average
value of the control group.
}
