% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{boxplotRelativeGeneExpression}
\alias{boxplotRelativeGeneExpression}
\title{boxplotRelativeGeneExpression}
\usage{
boxplotRelativeGeneExpression(geneDataFrame)
}
\arguments{
\item{geneDataFrame}{Dataframe produced by readGenes with averaged duplicates.}
}
\description{
This function plots relative expression by group
}
