% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filterDuplicates}
\alias{filterDuplicates}
\title{filterDuplicates}
\usage{
filterDuplicates(geneDataFrame, threshold = 2)
}
\arguments{
\item{geneDataFrame}{Dataframe produced by readGenes.}

\item{threshold}{Threshold for removing bad duplicates.}
}
\value{
geneDataFrame without bad duplicates.
}
\description{
This function takes a dataframe produced by readGenes.
Removes duplicates if ratio is greater than a given threshold.
}
