% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filterOutliers}
\alias{filterOutliers}
\title{filterOutliers}
\usage{
filterOutliers(relativeExpDF, groupCol, geneCol, expressionCol, gene,
  controlGroup, treatedGroup)
}
\arguments{
\item{relativeExpDF}{Dataframe with gene relative expression information.}

\item{groupCol}{Grouping column to be used for difference testing.}

\item{geneCol}{Column containing gene names.}

\item{expressionCol}{Column containing relative expression.}

\item{gene}{Gene to test.}

\item{controlGroup}{Name of the control group.}

\item{treatedGroup}{Name of the treated group.
("two.sided", "greater" or "less").}
}
\description{
This function removes outliers in two groups for a single gene within a
relativeExpression dataframe.
}
