% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hello}
\alias{hello}
\title{Hello function}
\usage{
hello()
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
hello()
}
\keyword{cats}
