% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotGene}
\alias{plotGene}
\title{plotGene}
\usage{
plotGene(relativeExpDF, groupCol, geneCol, expressionCol, gene,
  alternative = "two.sided", filterOutliers = T,
  controlRelativized = F, controlGroup, treatedGroup, maxY = 2)
}
\arguments{
\item{relativeExpDF}{Dataframe with gene relative expression information.}

\item{groupCol}{Grouping column to be used for difference testing.}

\item{geneCol}{Column containing gene names.}

\item{expressionCol}{Column containing relative expression.}

\item{gene}{Gene to test.}

\item{alternative}{Alternative hypothesis for the test
("two.sided", "greater" or "less").}

\item{filterOutliers}{Whether to remove outliers or not. Only one outlier per
group based on Grubb's test.}

\item{controlRelativized}{Display expression relativized to control group
(i.e. control group mean expression = 1)}

\item{controlGroup}{Name of the control group.}

\item{treatedGroup}{Name of the treated group.}

\item{maxY}{Upper limit of y axis.}
}
\description{
This function plots a bar and dots graph for a single gene
comparing two groups using the testGene function.
}
