% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotHouseKeepings}
\alias{plotHouseKeepings}
\title{plotHouseKeepings}
\usage{
plotHouseKeepings(geneDataFrame, houseKeepingNames)
}
\arguments{
\item{geneDataFrame}{Dataframe produced by readGenes with averaged duplicates.}

\item{houseKeepingNames}{Vector with names of genes to include in the plot.}
}
\description{
This function plots average N0 vs Sample for housekeepings diagnostics.
}
