% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readGenes}
\alias{readGenes}
\title{readGenes}
\usage{
readGenes(excel)
}
\arguments{
\item{excel}{Path of the excel file to read.}
}
\value{
A dataframe with data read from the excel file.
}
\description{
This function reads gene data from an excel file.
Por ahora tiene que tener las columnas: Well, Raton, Tratamiento, N0
}
