% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{testGene}
\alias{testGene}
\title{testGene}
\usage{
testGene(relativeExpDF, groupCol, geneCol, expressionCol, gene,
  alternative = "two.sided")
}
\arguments{
\item{relativeExpDF}{Dataframe with gene relative expression information.}

\item{groupCol}{Grouping column to be used for difference testing.}

\item{geneCol}{Column containing gene names.}

\item{expressionCol}{Column containing relative expression.}

\item{gene}{Gene to test.}

\item{alternative}{Alternative hypothesis for the test
("two.sided", "greater" or "less").}
}
\description{
This function test for significant differences in expression between
two groups for a single gene within a relativeExpression dataframe.
}
